refactor(levels/leaderboard): reworked leaderboard
- Now uses the new Levels service.
